:recycle: refactoring: Update deletarPostagem method to throw RuntimeException for not found Postagem instead of using ResponseStatusException

+++ src/main/java/com/conectados/conectados/services/PostagemService.java
@@ -36,6 +36,10 @@ public class PostagemService {
     }
 
     public void deletarPostagem(Long id) {
+        Optional<Postagem> postagem = buscarPostagemPorId(id);
+        if (postagem.isEmpty()) {
+            throw new RuntimeException("Postagem não encontrada");
+        }
         postagemRepository.deleteById(id);
     }
